- Moved the rest of the classes into the utilities package
