+++ src/burp/model/param/ParamGuesser.java
@@ -5,19 +5,19 @@ import burp.IHttpService;
 import burp.IParameter;
 import burp.IResponseVariations;
 import burp.IScannerInsertionPoint;
-import burp.Keysmith;
-import burp.Probe;
-import burp.WordProvider;
-import burp.albinowaxUtils.Attack;
+import burp.model.utilities.Keysmith;
+import burp.model.utilities.Probe;
+import burp.model.utilities.WordProvider;
+import burp.model.utilities.Attack;
 import burp.model.header.HeaderMutationGuesser;
 import burp.model.insertionPoints.ParamNameInsertionPoint;
 import burp.model.scanning.BulkScan;
 import burp.model.utilities.Utilities;
 import burp.view.ConfigurableSettings;
-import burp.albinowaxUtils.CustomScanIssue;
+import burp.model.utilities.CustomScanIssue;
 import burp.model.insertionPoints.ParamInsertionPoint;
 import burp.model.scanning.ParamScan;
-import burp.albinowaxUtils.PayloadInjector;
+import burp.model.utilities.PayloadInjector;
 import burp.model.insertionPoints.RawInsertionPoint;
 import burp.model.scanning.Scan;
 import org.apache.commons.collections4.queue.CircularFifoQueue;
