+++ src/burp/model/scanning/RailsUtmScan.java
@@ -5,7 +5,7 @@ import burp.IHttpRequestResponse;
 import burp.IHttpService;
 import burp.IScanIssue;
 import burp.IScannerInsertionPoint;
-import burp.albinowaxUtils.Resp;
+import burp.model.utilities.Resp;
 import burp.model.utilities.Utilities;
 
 import java.util.List;
