+++ src/burp/model/scanning/Scan.java
@@ -15,10 +15,10 @@ import burp.IHttpService;
 import burp.IScanIssue;
 import burp.IScannerCheck;
 import burp.IScannerInsertionPoint;
-import burp.albinowaxUtils.CustomScanIssue;
-import burp.albinowaxUtils.Req;
-import burp.albinowaxUtils.Resp;
-import burp.albinowaxUtils.ZgrabLoader;
+import burp.model.utilities.CustomScanIssue;
+import burp.model.utilities.Req;
+import burp.model.utilities.Resp;
+import burp.model.utilities.ZgrabLoader;
 import burp.model.utilities.Utilities;
 import burp.view.SettingsBox;
 import org.apache.commons.lang3.NotImplementedException;
