+++ src/burp/model/utilities/Fuzzable.java
@@ -1,8 +1,7 @@
-package burp.albinowaxUtils;
+package burp.model.utilities;
 
 import burp.IHttpRequestResponse;
 
-import burp.model.utilities.Utilities;
 import org.apache.commons.lang3.ArrayUtils;
 
 public class Fuzzable extends CustomScanIssue {
