+++ src/burp/model/utilities/PayloadInjector.java
@@ -3,14 +3,12 @@
 // (powered by FernFlower decompiler)
 //
 
-package burp.albinowaxUtils;
+package burp.model.utilities;
 
 import burp.IHttpRequestResponse;
 import burp.IHttpService;
 import burp.IScannerInsertionPoint;
-import burp.Probe;
 import burp.model.header.HeaderMutator;
-import burp.model.utilities.Utilities;
 
 import java.io.IOException;
 import java.util.ArrayList;
