+++ src/burp/model/utilities/Utilities.java
@@ -10,12 +10,7 @@ import burp.IResponseInfo;
 import burp.IScanIssue;
 import burp.IScannerInsertionPoint;
 import burp.model.param.PartialParam;
-import burp.Probe;
-import burp.albinowaxUtils.Attack;
 import burp.view.ConfigurableSettings;
-import burp.albinowaxUtils.Fuzzable;
-import burp.albinowaxUtils.LazyRequestInfo;
-import burp.albinowaxUtils.Resp;
 import org.apache.commons.collections4.queue.CircularFifoQueue;
 import org.apache.commons.lang3.CharUtils;
 import org.apache.commons.lang3.StringEscapeUtils;
